+++ src/main/java/com/kodilla/ecommercee/domain/dto/GroupDto.java
@@ -1,4 +1,4 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.domain.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
